Added decay equation, updated inputs, unit tests, ect.

+++ src/edu/mtu/simulation/schedule/Schedule.java
@@ -61,8 +61,10 @@ public class Schedule {
 	
 	/**
 	 * 
+	 * @param simulation
+	 * @param runTill
 	 */
-	public void start(Simulation simulation, int runTill) {
+	public void start(Simulation simulation, long runTill) {
 		// Check to make sure a simulation was provided
 		if (simulation == null) {
 			throw new IllegalArgumentException("The simulation cannot be null");
